modified the data type of ID to uint32 in the three structs

+++ data/std_lib.h
@@ -35,7 +35,7 @@ typedef float                 float32;
 typedef double                float64;
 
 typedef struct {
-    uint16 ID;
+    uint32 ID;
     uint8 name[NAME_SIZE];
     uint8 gender;
     uint8 age;
@@ -43,13 +43,13 @@ typedef struct {
 } Record;
 
 typedef struct {
-    uint16 ID;
+    uint32 ID;
     uint8 name[NAME_SIZE];
     uint8 password[PASSWORD_SIZE];
 } Admin;
 
 typedef struct {
-    uint16 ID;
+    uint32 ID;
     uint8 password[PASSWORD_SIZE];
 } Student;
 
@@ -70,5 +70,6 @@ typedef struct {
  *  Muhammad Wael          5/5/2024 21:41           Adding File Layout
  *  Muhammad Wael          5/5/2024 22:00           Adding typedef section
  *  Muhammad Wael          8/5/2024 22:05           modifying typedef section
+ *  Mina Nabil             9/5/2024 22:42           modified the data type of ID to uint32 in the three structs
  */
 /* ****************** History Log Section End ****************** */
